fix(terminal): honor readline line navigation across IMEs
Use physical keyboard codes for Ctrl+A and Ctrl+E so readline line-boundary controls remain available across Korean, Chinese, Japanese, and Russian input sources. Preserve Kitty CSI-u handling and leave active IME composition untouched. Add focused multilingual and composition regression coverage.
